Move some attributes from AstroDataFits to the base class
